Move ObjectServer, ProxyService to their own files.
