+++ lib/dbus.rb
@@ -43,10 +43,12 @@ require_relative "dbus/node_tree"
 require_relative "dbus/object"
 require_relative "dbus/object_manager"
 require_relative "dbus/object_path"
+require_relative "dbus/object_server"
 require_relative "dbus/platform"
 require_relative "dbus/proxy_object"
 require_relative "dbus/proxy_object_factory"
 require_relative "dbus/proxy_object_interface"
+require_relative "dbus/proxy_service"
 require_relative "dbus/raw_message"
 require_relative "dbus/type"
 require_relative "dbus/xml"
