+++ lib/dbus/bus.rb
@@ -13,201 +13,10 @@
 require "socket"
 require "singleton"
 
-require_relative "node_tree"
-
 # = D-Bus main module
 #
 # Module containing all the D-Bus modules and classes.
 module DBus
-  # Used by clients to represent a named service on the other side of the bus.
-  #
-  # Formerly this class was intermixed with {ObjectServer} as Service.
-  #
-  # @example Usage
-  #   svc = DBus.system_bus["org.freedesktop.machine1"]
-  #   manager = svc["/org/freedesktop/machine1"]
-  #   p manager.ListImages
-  class ProxyService < NodeTree
-    # @return [BusName,nil] The service name.
-    # May be nil for peer connections
-    attr_reader :name
-    # @return [Connection] The connection we're using.
-    attr_reader :connection
-
-    def initialize(name, connection)
-      @name = BusName.new(name)
-      @connection = connection
-      super()
-    end
-
-    # Determine whether the service name already exists.
-    def exists?
-      bus = connection # TODO: raise a better error if this is a peer connection
-      bus.proxy.ListNames[0].member?(@name)
-    end
-
-    # Perform an introspection on all the objects on the service
-    # (starting recursively from the root).
-    def introspect
-      raise NotImplementedError if block_given?
-
-      rec_introspect(@root, "/")
-      self
-    end
-
-    # Retrieves an object at the given _path_.
-    # @param path [ObjectPath]
-    # @return [ProxyObject]
-    def [](path)
-      object(path, api: ApiOptions::A1)
-    end
-
-    # Retrieves an object at the given _path_
-    # whose methods always return an array.
-    # @param path [ObjectPath]
-    # @param api [ApiOptions]
-    # @return [ProxyObject]
-    def object(path, api: ApiOptions::A0)
-      node = get_node(path, create: true)
-      if node.object.nil? || node.object.api != api
-        node.object = ProxyObject.new(
-          @connection, @name, path,
-          api: api
-        )
-      end
-      node.object
-    end
-
-    private
-
-    # Perform a recursive retrospection on the given current _node_
-    # on the given _path_.
-    def rec_introspect(node, path)
-      xml = bus.introspect_data(@name, path)
-      intfs, subnodes = IntrospectXMLParser.new(xml).parse
-      subnodes.each do |nodename|
-        subnode = node[nodename] = Node.new(nodename)
-        subpath = if path == "/"
-                    "/#{nodename}"
-                  else
-                    "#{path}/#{nodename}"
-                  end
-        rec_introspect(subnode, subpath)
-      end
-      return if intfs.empty?
-
-      node.object = ProxyObjectFactory.new(xml, @connection, @name, path).build
-    end
-  end
-
-  # The part of a {Connection} that can export {DBus::Object}s to provide
-  # services to clients.
-  #
-  # Note that an ObjectServer does not have a name. Typically a {Connection}
-  # has one well known name, but can have none or more.
-  #
-  # Formerly this class was intermixed with {ProxyService} as Service.
-  #
-  # @example Usage
-  #   bus = DBus.session_bus
-  #   svr = bus.object_server
-  #   obj = DBus::Object.new("/path") # a subclass more likely
-  #   svr.export(obj)
-  #   bus.request_service("org.example.Test") # FIXME request_name
-  #   # FIXME: convenience, combine exporting root object with a name
-  class ObjectServer < NodeTree
-    # @return [Connection] The connection we're using.
-    attr_reader :connection
-
-    def initialize(connection)
-      @connection = connection
-      super()
-    end
-
-    # Retrieves an object at the given _path_
-    # @param path [ObjectPath]
-    # @return [DBus::Object]
-    def object(path)
-      node = get_node(path, create: false)
-      node&.object
-    end
-    alias [] object
-
-    # Export an object
-    # @param obj [DBus::Object]
-    def export(obj)
-      node = get_node(obj.path, create: true)
-      # TODO: clarify that this is indeed the right thing, and document
-      # raise "At #{obj.path} there is already an object #{node.object.inspect}" if node.object
-
-      node.object = obj
-
-      obj.connection = @connection
-      object_manager_for(obj)&.object_added(obj)
-    end
-
-    # Undo exporting an object _obj_.
-    # Raises ArgumentError if it is not a DBus::Object.
-    # Returns the object, or false if _obj_ was not exported.
-    # @param obj [DBus::Object]
-    def unexport(obj)
-      raise ArgumentError, "Expecting a DBus::Object argument" unless obj.is_a?(DBus::Object)
-
-      last_path_separator_idx = obj.path.rindex("/")
-      parent_path = obj.path[1..last_path_separator_idx - 1]
-      node_name = obj.path[last_path_separator_idx + 1..-1]
-
-      parent_node = get_node(parent_path, create: false)
-      return false unless parent_node
-
-      object_manager_for(obj)&.object_removed(obj)
-      obj.connection = nil
-      parent_node.delete(node_name).object
-    end
-
-    # Find the (closest) parent of *object*
-    # implementing the ObjectManager interface, or nil
-    # @return [DBus::Object,nil]
-    def object_manager_for(object)
-      path = object.path
-      node_chain = get_node_chain(path)
-      om_node = node_chain.reverse_each.find do |node|
-        node.object&.is_a? DBus::ObjectManager
-      end
-      om_node&.object
-    end
-
-    # All objects (not paths) under this path (except itself).
-    # @param path [ObjectPath]
-    # @return [Array<DBus::Object>]
-    # @raise ArgumentError if the *path* does not exist
-    def descendants_for(path)
-      node = get_node(path, create: false)
-      raise ArgumentError, "Object path #{path} doesn't exist" if node.nil?
-
-      node.descendant_objects
-    end
-
-    #########
-
-    private
-
-    #########
-
-    # @raise ArgumentError if the *path* does not exist
-    def get_node_chain(path)
-      n = @root
-      result = [n]
-      path.sub(%r{^/}, "").split("/").each do |elem|
-        n = n[elem]
-        raise ArgumentError, "Object path #{path} doesn't exist" if n.nil?
-
-        result.push(n)
-      end
-      result
-    end
-  end
-
   # FIXME: rename Connection to Bus?
 
   # D-Bus main connection class
